% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rdsdataservice_operations.R
\name{execute_sql}
\alias{execute_sql}
\title{Executes any SQL statement on the target database synchronously}
\usage{
execute_sql(
  awsSecretStoreArn = "string",
  database = "string",
  dbClusterOrInstanceArn = "string",
  schema = "string",
  sqlStatements = "string"
)
}
\arguments{
\item{awsSecretStoreArn}{[required] ARN of the db credentials in AWS Secret Store or the friendly secret name}

\item{database}{Target DB name}

\item{dbClusterOrInstanceArn}{[required] ARN of the target db cluster or instance}

\item{schema}{Target Schema name}

\item{sqlStatements}{[required] SQL statement(s) to be executed. Statements can be chained by using semicolons}
}
\description{
Executes any SQL statement on the target database synchronously
}
\examples{

}
