% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rdsdataservice_package.R
\docType{package}
\name{paws.rdsdataservice-package}
\alias{paws.rdsdataservice}
\alias{paws.rdsdataservice-package}
\title{paws.rdsdataservice: AWS RDS DataService}
\description{
AWS RDS DataService provides Http Endpoint to query
    RDS databases.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
